Set up expanded state for StudentCard
This commit adds expanded boolean state to the StudentCard component.
When this state is true, the card will display all of the student grades
When false, the grades will not be displayed.

Future commits will implement a button to toggle the expanded state
between true and false

For the moment, **there are no user facing changes** in this commit